Trust Slider moved to general level on shared
