+++ src/app/features/simulator/components/simulation-input-details/simulation-input-details.component.ts
@@ -6,7 +6,7 @@ import {SimulationInputs} from "@app/features/simulator/models/simulation-inputs
 import {NgForOf} from "@angular/common";
 import {ConversationPartner} from "@app/shared/keml/core/conversation-partner";
 import {TrustComputationService} from "@app/features/simulator/services/trust-computation.service";
-import {TrustSliderComponent} from "@app/shared/keml/graphical/helper/trust/trust-slider/trust-slider.component";
+import {TrustSliderComponent} from "@app/shared/trust-slider/trust-slider.component";
 
 @Component({
     selector: 'simulation-input-details',
