+++ src/app/shared/trust-slider/trust-slider.component.ts
